pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ops-risk-env"
version = "1.0.0"
description = "Operational Risk Incident Triage"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "openai",
    "requests",
    "openenv-core",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
py-modules = ["inference"]
packages = ["server", "ops_sim"]
